requirements.txt
tqdm
base64
tiktoken
torchvision
langdetect
transformers==4.32.0
transformers_stream_generator==0.0.4
